Changed WaitingLobbyComponents into components package with (ColorSeal, CountdownOverlay)
